Extend doorkeeper 3 with its extracted ORMs
doorkeeper 3 has not been released yet, and will be when its ORMs are
successfully extracted and tested into this project.

This PR extends doorkeeper with the ORM specifics for mongoid and
mongomapper.

To test it, it loads doorkeeper as a submodule, copy its test suite into
this project, configures the ORMs according to the matrix defined in
`.travis.yml` file, and run the test suite with each.

Renames project from `doorkeeper-orm` to `doorkeeper-mongodb`.